+++ core/class/enedis.class.php
@@ -21,24 +21,6 @@ require_once __DIR__  . '/../../../../core/php/core.inc.php';
 
 class enedis extends eqLogic {
 
-  public static function cron() {
-    $cronMinute = config::byKey('cronMinute', __CLASS__);
-    if (!empty($cronMinute) && date('i') != $cronMinute) return;
-
-    $eqLogics = self::byType(__CLASS__, true);
-
-    foreach ($eqLogics as $eqLogic) {
-      if (date('G') < 5 || date('G') >= 20) {
-        if ($eqLogic->getCache('getEnedisData') == 'done') {
-          $eqLogic->setCache('getEnedisData', null);
-        }
-      }
-      else if ($eqLogic->getCache('getEnedisData') != 'done') {
-        $eqLogic->refreshData();
-      }
-    }
-  }
-
   public static function dependancy_info() {
     $return = array();
     $return['progress_file'] = jeedom::getTmpFolder('enedis') . '/dependance';
@@ -46,6 +28,7 @@ class enedis extends eqLogic {
     if (exec("dpkg-query -W -f='\${Status}\n' php-mbstring") == 'unknown ok not-installed') {
       $return['state'] = 'nok';
     }
+    /* Since v4.2 */
     // $packages = system::checkAndInstall(json_decode(file_get_contents(__DIR__.'/../../plugin_info/packages.json'),true));
     // $return['state'] = ($packages['apt::php-mbstring']['status'] == 1) ? 'ok' : 'nok';
     return $return;
@@ -56,33 +39,67 @@ class enedis extends eqLogic {
     return array('script' => dirname(__FILE__) . '/../../resources/install_#stype#.sh ' . jeedom::getTmpFolder('enedis') . '/dependance', 'log' => log::getPathToLog(__CLASS__ . '_update'));
   }
 
-  public function refreshData($startDate = null) {
-    log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Interrogation des serveurs Enedis',__FILE__));
-    $usagePointId = $this->getConfiguration('usage_point_id');
-    $measureTypes = ($this->getConfiguration('measure_type') != 'both') ? [$this->getConfiguration('measure_type')] : array('consumption', 'production');
-    $start_date = (empty($startDate)) ? date('Y-m-d', strtotime('first day of January')) : $startDate;
-    $start_date_load = (empty($startDate)) ? date('Y-m-d', strtotime('-1 day')) : date('Y-m-d', strtotime('-7 days'));
-    $end_date = date('Y-m-d');
-    $need_refresh = false;
-
-    foreach ($measureTypes as $measureType) {
-      $dailyCmd = $this->getCmd('info', 'daily_'.$measureType);
-      $dailyCmd->execCmd();
-      if (empty($startDate) && $dailyCmd->getCollectDate() >= date('Y-m-d', strtotime('-1 day'))) {
-        log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $dailyCmd->getName() . __('] Données déjà enregistrées pour le ',__FILE__) . date('d/m/Y', strtotime('-1 day')));
+  public static function pull($_options) {
+    $eqLogic = enedis::byId($_options['enedis_id']);
+    if (!is_object($eqLogic)) {
+      $cron = cron::byClassAndFunction(__CLASS__, 'pull', $_options);
+      if (is_object($cron)) {
+        $cron->remove();
       }
-      else {
+      throw new Exception(__('Tâche supprimée car équipement non trouvé (ID) : ', __FILE__) . $_options['enedis_id']);
+    }
+    sleep(rand(1,59));
+    $eqLogic->refreshData();
+  }
+
+  public function reschedule($_next = null) {
+    if (empty($_next)) {
+      $_next = strtotime('+1 day ' . date('Y-m-d 07:'.rand(1,59)));
+    }
+    log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Prochaine programmation : ',__FILE__) . date('d/m/Y H:i', $_next));
+    $options = array('enedis_id' => intval($this->getId()));
+    $cron = cron::byClassAndFunction(__CLASS__, 'pull', $options);
+    if (is_object($cron)) {
+      $cron->remove(false);
+    }
+    $cron = (new cron)
+    ->setClass(__CLASS__)
+    ->setFunction('pull')
+    ->setOption($options)
+    ->setTimeout(1440)
+    ->setOnce(1);
+    $cron->setSchedule(cron::convertDateToCron($_next));
+    $cron->save();
+  }
+
+  public function refreshData($startDate = null) {
+    if ($this->getIsEnable() == 1) {
+      log::add(__CLASS__, 'debug', $this->getHumanName() .' -----------------------------------------------------------------------');
+      log::add(__CLASS__, 'debug', $this->getHumanName() . __(' *** Interrogation des serveurs Enedis ***',__FILE__));
+      $usagePointId = $this->getConfiguration('usage_point_id');
+      $measureTypes = ($this->getConfiguration('measure_type') != 'both') ? [$this->getConfiguration('measure_type')] : array('consumption', 'production');
+      $start_date = (empty($startDate)) ? date('Y-m-d', strtotime('first day of January')) : $startDate;
+      $start_date_load = (empty($startDate)) ? date('Y-m-d', strtotime('-1 day')) : date('Y-m-d', strtotime('-7 days'));
+      $end_date = date('Y-m-d');
+      $refresh_day = true;
+      $refresh_power = true;
+      $refresh_load = true;
+
+      foreach ($measureTypes as $measureType) {
+        $dailyCmd = $this->getCmd('info', 'daily_'.$measureType);
         $monthlyCmd = $this->getCmd('info', 'monthly_'.$measureType);
         $yearlyCmd = $this->getCmd('info', 'yearly_'.$measureType);
         $returnMonthValue = 0;
         $returnYearValue = 0;
-        $need_refresh = true;
         $data = $this->getData('/metering_data/daily_'.$measureType.'?start='.$start_date.'&end='.$end_date.'&usage_point_id='.$usagePointId);
         if(isset($data['meter_reading']) && isset($data['meter_reading']['interval_reading'])) {
           foreach ($data['meter_reading']['interval_reading'] as $value) {
             $valueTimestamp = strtotime($value['date']);
             $this->checkData($dailyCmd, $value['value'], date('Y-m-d 00:00:00', $valueTimestamp));
 
+            if (date('Y-m-d', $valueTimestamp) >= date('Y-m-d', strtotime('-1 day'))) {
+              $refresh_day = false;
+            }
             if ($value['date'] == date('Y-m-01', $valueTimestamp)) {
               $returnMonthValue = $value['value'];
             }
@@ -103,48 +120,49 @@ class enedis extends eqLogic {
         else if (isset($data['error'])) {
           log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $dailyCmd->getName() . __('] Erreur sur la récupération des données : ',__FILE__) . $data['error'] . ' ' . $data['error_description']);
         }
-      }
 
-      $dailyMaxCmd = $this->getCmd('info', 'daily_'.$measureType.'_max_power');
-      $dailyMaxCmd->execCmd();
-      if (empty($startDate) && $dailyMaxCmd->getCollectDate() >= date('Y-m-d', strtotime('-1 day'))) {
-        log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $dailyMaxCmd->getName() . __('] Données déjà enregistrées pour le ',__FILE__) . date('d/m/Y', strtotime('-1 day')));
-      }
-      else {
-        $need_refresh = true;
+        $dailyMaxCmd = $this->getCmd('info', 'daily_'.$measureType.'_max_power');
         $data = $this->getData('/metering_data/daily_'.$measureType.'_max_power?start='.$start_date.'&end='.$end_date.'&usage_point_id='.$usagePointId);
         if(isset($data['meter_reading']) && isset($data['meter_reading']['interval_reading'])){
           foreach ($data['meter_reading']['interval_reading'] as $value) {
             $this->checkData($dailyMaxCmd, $value['value'], $value['date']);
+            if ($value['date'] >= date('Y-m-d', strtotime('-1 day'))) {
+              $refresh_power = false;
+            }
           }
         }
         else if (isset($data['error'])) {
           log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $dailyMaxCmd->getName() . __('] Erreur sur la récupération des données : ',__FILE__) . $data['error'] . ' ' . $data['error_description']);
         }
-      }
 
-      $loadCmd = $this->getCmd('info', $measureType.'_load_curve');
-      $loadCmd->execCmd();
-      if ($loadCmd->getCollectDate() >= date('Y-m-d', strtotime('today'))) {
-        log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $loadCmd->getName() . __('] Données déjà enregistrées pour le ',__FILE__) . date('d/m/Y', strtotime('-1 day')));
-      }
-      else {
-        $need_refresh = true;
+        $loadCmd = $this->getCmd('info', $measureType.'_load_curve');
         $data = $this->getData('/metering_data/'.$measureType.'_load_curve?start='.$start_date_load.'&end='.$end_date.'&usage_point_id='.$usagePointId);
         if (isset($data['meter_reading']) && isset($data['meter_reading']['interval_reading'])) {
           foreach ($data['meter_reading']['interval_reading'] as $value) {
             $this->checkData($loadCmd, $value['value'], $value['date']);
+            if ($value['date'] >= date('Y-m-d')) {
+              $refresh_load = false;
+            }
           }
         }
         else if (isset($data['error'])) {
           log::add(__CLASS__, 'debug', $this->getHumanName() . '[' . $loadCmd->getName() . __('] Erreur sur la récupération des données : ',__FILE__) . $data['error'] . ' ' . $data['error_description']);
         }
       }
-    }
 
-    if ($need_refresh === false && $this->getCache('getEnedisData') != 'done') {
-      $this->setCache('getEnedisData', 'done');
-      log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Toutes les données sont à jour - désactivation de la vérification automatique pour aujourd\'hui',__FILE__));
+      if (!$refresh_day && !$refresh_power && !$refresh_load) {
+        log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Toutes les données sont à jour',__FILE__));
+        $this->reschedule();
+      }
+      else if (date('H') >= 19) {
+        log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Fin des interrogations pour aujourd\'hui',__FILE__));
+        $this->reschedule();
+      }
+      else {
+        $next_launch = strtotime('+1 hour ' . date('Y-m-d H:i'));
+        log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Certaines données n\'ont pas été récupérées',__FILE__));
+        $this->reschedule($next_launch);
+      }
     }
   }
 
@@ -159,11 +177,7 @@ class enedis extends eqLogic {
   public function checkData($cmd, $value, $date) {
     if (strlen($date) === 19) {
       $rounded = new DateTime($date);
-      $rounded->setTime(
-        $rounded->format('H'),
-        floor($rounded->format('i') / 5) * 5,
-        0
-      );
+      $rounded->setTime($rounded->format('H'), floor($rounded->format('i') / 5) * 5, 0);
       $date = $rounded->format("Y-m-d H:i:s");
     }
 
@@ -195,7 +209,13 @@ class enedis extends eqLogic {
   }
 
   public function postUpdate() {
+    $options = array('enedis_id' => intval($this->getId()));
+    $cron = cron::byClassAndFunction(__CLASS__, 'pull', $options);
     if ($this->getIsEnable() == 1) {
+      if (!is_object($cron)) {
+        $this->reschedule();
+      }
+
       $refreshCmd = $this->getCmd(null, 'refresh');
       if (!is_object($refreshCmd)) {
         log::add(__CLASS__, 'debug', $this->getHumanName() . __(' Création commande : refresh/Rafraîchir',__FILE__));
@@ -210,7 +230,7 @@ class enedis extends eqLogic {
       }
 
       if (!is_file(dirname(__FILE__) . '/../../data/cmds/commands.json')) {
-        log::add(__CLASS__, 'debug', $this->getHumanName() . __('Fichier de création de commandes non trouvé',__FILE__));
+        log::add(__CLASS__, 'error', $this->getHumanName() . __('Fichier de création de commandes non trouvé',__FILE__));
       }
       else {
         $cmdsJson = file_get_contents(dirname(__FILE__) . '/../../data/cmds/commands.json');
@@ -224,12 +244,14 @@ class enedis extends eqLogic {
             $this->refreshData(date('Y-m-d', strtotime('-3 years')));
           }
         }
-        else if ($this->createCommands($cmdsArray[$measureType])){
+        else if ($this->createCommands($cmdsArray[$measureType])) {
           $this->refreshData(date('Y-m-d', strtotime('-3 years')));
         }
       }
     }
-
+    else if (is_object($cron)) {
+      $cron->remove(false);
+    }
   }
 
   public function createCommands($type) {